Completed writing tests for the prediction part.
Included bugfix for prediction error: "raise KeyError(f"{not_found} not in index")" when trying to predict on small sample sizes.

Updated model_abstract.py to more efficiently work with the getter and setter for the model features.

Updated test_prediction.py to remove some duplicated code. Updated the integration test for preprocessing to check for the presence of all model features.