fix: Loglevels fix
Cause: When logging.getLogger(name) gets called and a name is supplied,
logging creates a new logging instance from the root logger. The log
level changed for this new instance does not change the log level for
the root logger, causing problems when trying to set a different log
level from WARNING.]
Solution: Don't supply logging.getLogger with a
name and set the name manually. Now a root logger instance is created
with the correct log level.