fix: Imputer training clash with tests
### Changes

- Removed the `train=False` argument for imputer due to
clashes with tests. (Problem originated from the imputer exporting the
annotation features to `CapiceManager()`, but were then expected in the
test itself, while the input data can or can not contain the feature
itself).

### Notes

Due to the removal of the deprecated function that
would add a column if it was missing from the impute value keys (in the
json), CAPICE will ERROR at predicting stage due to missing features.