feat: variant Consequence extended
- Added the variant Consequence VEP processor to boolean True (1)
or
False (0) whenever a variant is considered to be one of
the
consequences. (example: if a variant is both frame_shift and
stop_lost,
both is_frame_shift and is_stop_lost will be set to 1).
-
Changed
version within setup.py to import the project version from
the
__version__.py file. (might potentially break something, since it
import
something from the project which it has to install in the
first
place).
- Bumped version to 3.0.0-beta2.

- A whole bunch of
deprecation warnings started to pop up with origin to numpy, pandas and
packaging.